Update scaleHost webhook to use hostTemplate
Change logic of the scaleHost webhook to use a hostTemplate as the
basis for creating new hosts. The old logic looked for existing hosts
with a label specified in the webhook. But with the introduction of
hostTemplates, which maintain all the config data necessary for
creating a new host, we no longer need to rely on copying data from
existing hosts.

We are maintaining backwards compatibility with previously created
webhooks that used the label-based approach.